+++ src/features/store/analytics/DeviceProductsActivity.tsx
